Added multiple container widgets
This is starting to shape up, multiple container widgets with content in them to help give an idea of what the final version might look like. This update has a lot of fucking around with sizing, resizing, positioning, and alignment. The biggest hurdle for right now is getting all this done up correctly.